small datatypes in Common Functions

+++ common/CommonFunctions.c
@@ -1,10 +1,10 @@
 #include "CommonFunctions.h"
 #include "ErrorCodes.h"
 
-unsigned long searchSumArrayLong(const unsigned long* const restrict array, unsigned long highestIndex, const unsigned long target) {
-    unsigned long lowestIndex = 0;
+uint_least32_t searchSumArrayLong(const uint_least32_t* const restrict array, uint_least32_t highestIndex, const uint_least32_t target) {
+    uint_least32_t lowestIndex = 0;
     do {
-        const unsigned long middleIndex = (lowestIndex + highestIndex) / 2;
+        const uint_least32_t middleIndex = (lowestIndex + highestIndex) / 2;
         if (target >= array[middleIndex]) {
             lowestIndex = middleIndex;
         }
@@ -15,10 +15,10 @@ unsigned long searchSumArrayLong(const unsigned long* const restrict array, unsi
     return (target >= array[highestIndex]) ? highestIndex : lowestIndex;
 }
 
-unsigned long searchSumArrayShort(const unsigned short* const restrict array, unsigned long highestIndex, const unsigned short target) {
-    unsigned long lowestIndex = 0;
+uint_least32_t searchSumArrayShort(const uint_least16_t* const restrict array, uint_least32_t highestIndex, const uint_least16_t target) {
+    uint_least32_t lowestIndex = 0;
     do {
-        const unsigned long middleIndex = (lowestIndex + highestIndex) / 2;
+        const uint_least32_t middleIndex = (lowestIndex + highestIndex) / 2;
         if (target >= array[middleIndex]) {
             lowestIndex = middleIndex;
         }
